requirements.txt
# requirements.txt
slack-bolt>=1.18.0
dbos>=0.9.0
gitpython>=3.1.40
pydantic>=2.0.0
httpx>=0.25.0
pytest>=7.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
asyncio>=3.4.3
pydantic-settings>=2.0.0
openai